L| Add fake-hwclock module for lidjamoypi
Hopefully this will cause the clock to be approximately restored
MUCH faster during boot than waiting for `ntp` to run after the
network is up.  Restoring the date&time is now part of the
activation script.

This is useful because without it, after a poweroff condition, the
lack of hwclock gives the system an ancient date/time, causing
`systemd`'s logging to be all kinds of not-helpful and incorrectly
garbage collected.