Add pibuilder user and authorizations
This adds a generally unprivileged user that the builder machine
can connect as when pushing new system builds to `lidjamoypi`.

The handful of privileged commands that are needed for this
purpose are explicitly allowed as a `NOPASSWD` `sudoers`
command set.

For posterity: The syntax and semantics of regex matched `sudoers`
entries is obscene!  You can't regex-match just the command arguments,
for instance.  It is all or nothing!